Extract HttpClientFailure into http_client_failure.dart
