+++ lib/src/sdk/errors.dart
@@ -1,12 +1,5 @@
 import 'failure.dart';
 
-class HttpClientFailure extends Failure {
-  HttpClientFailure({
-    required String message,
-    Map<String, String>? context,
-  }) : super(message: message, context: context);
-}
-
 class ResourceNotFoundFailure extends Failure {
   ResourceNotFoundFailure({
     required String resource,
