+++ lib/src/sdk/http_client.dart
@@ -4,9 +4,10 @@ import 'package:dio/dio.dart'
     show Dio, DioError, Options, Response, ResponseType;
 import 'package:trello_sdk/external/files.dart' show defaultFileWriter;
 
-import 'errors.dart' show HttpClientFailure, ResourceNotFoundFailure;
-import 'file_name.dart' show FileName;
-import 'query_options.dart' show QueryOptions;
+import 'errors.dart' show ResourceNotFoundFailure;
+import 'file_name.dart';
+import 'http_client_failure.dart';
+import 'query_options.dart';
 
 abstract class HttpResponse<T> {
   T? get data;
