+++ lib/src/sdk/http_client_failure.dart
@@ -1,0 +1,8 @@
+import 'failure.dart';
+
+class HttpClientFailure extends Failure {
+  HttpClientFailure({
+    required String message,
+    Map<String, String>? context,
+  }) : super(message: message, context: context);
+}
